feat(chrome): full-overlay desktop
The docked left sidebar becomes a floating panel over a full-bleed map, opened
from a search pill in the top-left corner or a right-edge control column
(layers, background map, alert bell with an in-view count). The background
picker slides out beside the column instead of hiding in a menu button.

hide_sidebar dies with the dock: whether the panel is open is runtime state, not
a setting. Old settings files ignore the removed field. The bottom timeline bar
stays docked until the scrubber lands.
